Auth controller

+++ src/main/kotlin/unq/pds/api/controller/CommissionController.kt
@@ -1,9 +1,15 @@
 package unq.pds.api.controller
 
-import io.swagger.v3.oas.annotations.*
-import io.swagger.v3.oas.annotations.media.*
-import io.swagger.v3.oas.annotations.responses.*
-import org.springframework.http.*
+import io.jsonwebtoken.Jwts
+import io.swagger.v3.oas.annotations.Operation
+import io.swagger.v3.oas.annotations.media.ArraySchema
+import io.swagger.v3.oas.annotations.media.Content
+import io.swagger.v3.oas.annotations.media.ExampleObject
+import io.swagger.v3.oas.annotations.media.Schema
+import io.swagger.v3.oas.annotations.responses.ApiResponse
+import io.swagger.v3.oas.annotations.responses.ApiResponses
+import org.springframework.http.HttpStatus
+import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.*
 import unq.pds.api.dtos.CommissionDTO
 import unq.pds.api.dtos.MessageDTO
@@ -57,8 +63,19 @@ class CommissionController(private val commissionService: CommissionService) {
                 )]
             )]
     )
-    fun createCommission(@RequestBody @Valid commission: CommissionDTO): ResponseEntity<Any> {
-        return ResponseEntity(commissionService.save(commission.fromDTOToModel()), HttpStatus.OK)
+    fun createCommission(
+        @CookieValue("jwt") jwt: String?,
+        @RequestBody @Valid commission: CommissionDTO
+    ): ResponseEntity<Any> {
+        if (jwt.isNullOrBlank()) {
+            return ResponseEntity(MessageDTO("It is not authenticated. Please log in"), HttpStatus.UNAUTHORIZED)
+        }
+        val body = Jwts.parser().setSigningKey("secret".encodeToByteArray()).parseClaimsJws(jwt).body
+        return if (body["role"] == "STUDENT") ResponseEntity(
+            MessageDTO("You do not have permissions to access this resource"),
+            HttpStatus.UNAUTHORIZED
+        )
+        else ResponseEntity(commissionService.save(commission.fromDTOToModel()), HttpStatus.OK)
     }
 
     @GetMapping
@@ -101,7 +118,10 @@ class CommissionController(private val commissionService: CommissionService) {
                 )]
             )]
     )
-    fun getCommission(@NotBlank @RequestParam id: Long): ResponseEntity<Any> {
+    fun getCommission(@CookieValue("jwt") jwt: String?, @NotBlank @RequestParam id: Long): ResponseEntity<Any> {
+        if (jwt.isNullOrBlank()) {
+            return ResponseEntity(MessageDTO("It is not authenticated. Please log in"), HttpStatus.UNAUTHORIZED)
+        }
         return ResponseEntity(commissionService.read(id), HttpStatus.OK)
     }
 
@@ -145,8 +165,16 @@ class CommissionController(private val commissionService: CommissionService) {
                 )]
             )]
     )
-    fun updateCommission(@RequestBody commission: Commission): ResponseEntity<Any> {
-        return ResponseEntity(commissionService.update(commission), HttpStatus.OK)
+    fun updateCommission(@CookieValue("jwt") jwt: String?, @RequestBody commission: Commission): ResponseEntity<Any> {
+        if (jwt.isNullOrBlank()) {
+            return ResponseEntity(MessageDTO("It is not authenticated. Please log in"), HttpStatus.UNAUTHORIZED)
+        }
+        val body = Jwts.parser().setSigningKey("secret".encodeToByteArray()).parseClaimsJws(jwt).body
+        return if (body["role"] == "STUDENT") ResponseEntity(
+            MessageDTO("You do not have permissions to access this resource"),
+            HttpStatus.UNAUTHORIZED
+        )
+        else ResponseEntity(commissionService.update(commission), HttpStatus.OK)
     }
 
     @DeleteMapping
@@ -191,7 +219,15 @@ class CommissionController(private val commissionService: CommissionService) {
                 )]
             )]
     )
-    fun deleteCommission(@NotBlank @RequestParam id: Long): ResponseEntity<Any> {
+    fun deleteCommission(@CookieValue("jwt") jwt: String?, @NotBlank @RequestParam id: Long): ResponseEntity<Any> {
+        if (jwt.isNullOrBlank()) {
+            return ResponseEntity(MessageDTO("It is not authenticated. Please log in"), HttpStatus.UNAUTHORIZED)
+        }
+        val body = Jwts.parser().setSigningKey("secret".encodeToByteArray()).parseClaimsJws(jwt).body
+        if (body["role"] == "STUDENT") return ResponseEntity(
+            MessageDTO("You do not have permissions to access this resource"),
+            HttpStatus.UNAUTHORIZED
+        )
         commissionService.delete(id)
         return ResponseEntity(MessageDTO("Commission has been deleted successfully"), HttpStatus.OK)
     }
@@ -236,8 +272,20 @@ class CommissionController(private val commissionService: CommissionService) {
                 )]
             )]
     )
-    fun addStudent(@NotBlank @PathVariable commissionId: Long, @NotBlank @PathVariable studentId: Long): ResponseEntity<Any> {
-        return ResponseEntity(commissionService.addStudent(commissionId, studentId), HttpStatus.OK)
+    fun addStudent(
+        @CookieValue("jwt") jwt: String?,
+        @NotBlank @PathVariable commissionId: Long,
+        @NotBlank @PathVariable studentId: Long
+    ): ResponseEntity<Any> {
+        if (jwt.isNullOrBlank()) {
+            return ResponseEntity(MessageDTO("It is not authenticated. Please log in"), HttpStatus.UNAUTHORIZED)
+        }
+        val body = Jwts.parser().setSigningKey("secret".encodeToByteArray()).parseClaimsJws(jwt).body
+        return if (body["role"] == "STUDENT") ResponseEntity(
+            MessageDTO("You do not have permissions to access this resource"),
+            HttpStatus.UNAUTHORIZED
+        )
+        else ResponseEntity(commissionService.addStudent(commissionId, studentId), HttpStatus.OK)
     }
 
     @PutMapping("/removeStudent/{commissionId}/{studentId}")
@@ -280,8 +328,20 @@ class CommissionController(private val commissionService: CommissionService) {
                 )]
             )]
     )
-    fun removeStudent(@NotBlank @PathVariable commissionId: Long, @NotBlank @PathVariable studentId: Long): ResponseEntity<Any> {
-        return ResponseEntity(commissionService.removeStudent(commissionId, studentId), HttpStatus.OK)
+    fun removeStudent(
+        @CookieValue("jwt") jwt: String?,
+        @NotBlank @PathVariable commissionId: Long,
+        @NotBlank @PathVariable studentId: Long
+    ): ResponseEntity<Any> {
+        if (jwt.isNullOrBlank()) {
+            return ResponseEntity(MessageDTO("It is not authenticated. Please log in"), HttpStatus.UNAUTHORIZED)
+        }
+        val body = Jwts.parser().setSigningKey("secret".encodeToByteArray()).parseClaimsJws(jwt).body
+        return if (body["role"] == "STUDENT") ResponseEntity(
+            MessageDTO("You do not have permissions to access this resource"),
+            HttpStatus.UNAUTHORIZED
+        )
+        else ResponseEntity(commissionService.removeStudent(commissionId, studentId), HttpStatus.OK)
     }
 
     @PutMapping("/addTeacher/{commissionId}/{teacherId}")
@@ -324,8 +384,20 @@ class CommissionController(private val commissionService: CommissionService) {
                 )]
             )]
     )
-    fun addTeacher(@NotBlank @PathVariable commissionId: Long, @NotBlank @PathVariable teacherId: Long): ResponseEntity<Any> {
-        return ResponseEntity(commissionService.addTeacher(commissionId, teacherId), HttpStatus.OK)
+    fun addTeacher(
+        @CookieValue("jwt") jwt: String?,
+        @NotBlank @PathVariable commissionId: Long,
+        @NotBlank @PathVariable teacherId: Long
+    ): ResponseEntity<Any> {
+        if (jwt.isNullOrBlank()) {
+            return ResponseEntity(MessageDTO("It is not authenticated. Please log in"), HttpStatus.UNAUTHORIZED)
+        }
+        val body = Jwts.parser().setSigningKey("secret".encodeToByteArray()).parseClaimsJws(jwt).body
+        return if (body["role"] == "STUDENT") ResponseEntity(
+            MessageDTO("You do not have permissions to access this resource"),
+            HttpStatus.UNAUTHORIZED
+        )
+        else ResponseEntity(commissionService.addTeacher(commissionId, teacherId), HttpStatus.OK)
     }
 
     @PutMapping("/removeTeacher/{commissionId}/{teacherId}")
@@ -368,8 +440,20 @@ class CommissionController(private val commissionService: CommissionService) {
                 )]
             )]
     )
-    fun removeTeacher(@NotBlank @PathVariable commissionId: Long, @NotBlank @PathVariable teacherId: Long): ResponseEntity<Any> {
-        return ResponseEntity(commissionService.removeTeacher(commissionId, teacherId), HttpStatus.OK)
+    fun removeTeacher(
+        @CookieValue("jwt") jwt: String?,
+        @NotBlank @PathVariable commissionId: Long,
+        @NotBlank @PathVariable teacherId: Long
+    ): ResponseEntity<Any> {
+        if (jwt.isNullOrBlank()) {
+            return ResponseEntity(MessageDTO("It is not authenticated. Please log in"), HttpStatus.UNAUTHORIZED)
+        }
+        val body = Jwts.parser().setSigningKey("secret".encodeToByteArray()).parseClaimsJws(jwt).body
+        return if (body["role"] == "STUDENT") ResponseEntity(
+            MessageDTO("You do not have permissions to access this resource"),
+            HttpStatus.UNAUTHORIZED
+        )
+        else ResponseEntity(commissionService.removeTeacher(commissionId, teacherId), HttpStatus.OK)
     }
 
     @PutMapping("/addGroup/{commissionId}/{groupId}")
@@ -412,8 +496,20 @@ class CommissionController(private val commissionService: CommissionService) {
                 )]
             )]
     )
-    fun addGroup(@NotBlank @PathVariable commissionId: Long, @NotBlank @PathVariable groupId: Long): ResponseEntity<Any> {
-        return ResponseEntity(commissionService.addGroup(commissionId, groupId), HttpStatus.OK)
+    fun addGroup(
+        @CookieValue("jwt") jwt: String?,
+        @NotBlank @PathVariable commissionId: Long,
+        @NotBlank @PathVariable groupId: Long
+    ): ResponseEntity<Any> {
+        if (jwt.isNullOrBlank()) {
+            return ResponseEntity(MessageDTO("It is not authenticated. Please log in"), HttpStatus.UNAUTHORIZED)
+        }
+        val body = Jwts.parser().setSigningKey("secret".encodeToByteArray()).parseClaimsJws(jwt).body
+        return if (body["role"] == "STUDENT") ResponseEntity(
+            MessageDTO("You do not have permissions to access this resource"),
+            HttpStatus.UNAUTHORIZED
+        )
+        else ResponseEntity(commissionService.addGroup(commissionId, groupId), HttpStatus.OK)
     }
 
     @PutMapping("/removeGroup/{commissionId}/{groupId}")
@@ -456,8 +552,20 @@ class CommissionController(private val commissionService: CommissionService) {
                 )]
             )]
     )
-    fun removeGroup(@NotBlank @PathVariable commissionId: Long, @NotBlank @PathVariable groupId: Long): ResponseEntity<Any> {
-        return ResponseEntity(commissionService.removeGroup(commissionId, groupId), HttpStatus.OK)
+    fun removeGroup(
+        @CookieValue("jwt") jwt: String?,
+        @NotBlank @PathVariable commissionId: Long,
+        @NotBlank @PathVariable groupId: Long
+    ): ResponseEntity<Any> {
+        if (jwt.isNullOrBlank()) {
+            return ResponseEntity(MessageDTO("It is not authenticated. Please log in"), HttpStatus.UNAUTHORIZED)
+        }
+        val body = Jwts.parser().setSigningKey("secret".encodeToByteArray()).parseClaimsJws(jwt).body
+        return if (body["role"] == "STUDENT") ResponseEntity(
+            MessageDTO("You do not have permissions to access this resource"),
+            HttpStatus.UNAUTHORIZED
+        )
+        else ResponseEntity(commissionService.removeGroup(commissionId, groupId), HttpStatus.OK)
     }
 
     @GetMapping("/getAll")
@@ -478,7 +586,10 @@ class CommissionController(private val commissionService: CommissionService) {
                 ]
             )]
     )
-    fun getAll(): ResponseEntity<List<Commission>> {
+    fun getAll(@CookieValue("jwt") jwt: String?): ResponseEntity<Any> {
+        if (jwt.isNullOrBlank()) {
+            return ResponseEntity(MessageDTO("It is not authenticated. Please log in"), HttpStatus.UNAUTHORIZED)
+        }
         return ResponseEntity(commissionService.readAll(), HttpStatus.OK)
     }
 }
