+++ src/main/kotlin/unq/pds/api/controller/RepositoryController.kt
@@ -1,9 +1,15 @@
 package unq.pds.api.controller
 
-import io.swagger.v3.oas.annotations.*
-import io.swagger.v3.oas.annotations.media.*
-import io.swagger.v3.oas.annotations.responses.*
-import org.springframework.http.*
+import io.jsonwebtoken.Jwts
+import io.swagger.v3.oas.annotations.Operation
+import io.swagger.v3.oas.annotations.media.ArraySchema
+import io.swagger.v3.oas.annotations.media.Content
+import io.swagger.v3.oas.annotations.media.ExampleObject
+import io.swagger.v3.oas.annotations.media.Schema
+import io.swagger.v3.oas.annotations.responses.ApiResponse
+import io.swagger.v3.oas.annotations.responses.ApiResponses
+import org.springframework.http.HttpStatus
+import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.*
 import unq.pds.api.dtos.MessageDTO
 import unq.pds.api.dtos.RepositoryDTO
@@ -46,7 +52,13 @@ class RepositoryController(private val repositoryService: RepositoryService) {
                 )]
             )]
     )
-    fun createRepository(@RequestBody @Valid repository: RepositoryDTO): ResponseEntity<Repository> {
+    fun createRepository(
+        @CookieValue("jwt") jwt: String?,
+        @RequestBody @Valid repository: RepositoryDTO
+    ): ResponseEntity<Any> {
+        if (jwt.isNullOrBlank()) {
+            return ResponseEntity(MessageDTO("It is not authenticated. Please log in"), HttpStatus.UNAUTHORIZED)
+        }
         return ResponseEntity(repositoryService.save(repository), HttpStatus.OK)
     }
 
@@ -90,8 +102,11 @@ class RepositoryController(private val repositoryService: RepositoryService) {
                 )]
             )]
     )
-    fun getRepository(@NotBlank @RequestParam id: Long): ResponseEntity<Any> {
-        return ResponseEntity(repositoryService.findById(id), HttpStatus.OK)
+    fun getRepository(@CookieValue("jwt") jwt: String?, @NotBlank @RequestParam id: Long): ResponseEntity<Any> {
+        if (jwt.isNullOrBlank()) {
+            return ResponseEntity(MessageDTO("It is not authenticated. Please log in"), HttpStatus.UNAUTHORIZED)
+        }
+        return  ResponseEntity(repositoryService.findById(id), HttpStatus.OK)
     }
 
     @PutMapping
@@ -134,8 +149,19 @@ class RepositoryController(private val repositoryService: RepositoryService) {
                 )]
             )]
     )
-    fun updateRepository(@RequestBody @Valid repository: RepositoryDTO): ResponseEntity<Any> {
-        return ResponseEntity(repositoryService.update(repository), HttpStatus.OK)
+    fun updateRepository(
+        @CookieValue("jwt") jwt: String?,
+        @RequestBody @Valid repository: RepositoryDTO
+    ): ResponseEntity<Any> {
+        if (jwt.isNullOrBlank()) {
+            return ResponseEntity(MessageDTO("It is not authenticated. Please log in"), HttpStatus.UNAUTHORIZED)
+        }
+        val body = Jwts.parser().setSigningKey("secret".encodeToByteArray()).parseClaimsJws(jwt).body
+        return if (body["role"] == "STUDENT") ResponseEntity(
+            MessageDTO("You do not have permissions to access this resource"),
+            HttpStatus.UNAUTHORIZED
+        )
+        else ResponseEntity(repositoryService.update(repository), HttpStatus.OK)
     }
 
     @DeleteMapping
@@ -180,7 +206,15 @@ class RepositoryController(private val repositoryService: RepositoryService) {
                 )]
             )]
     )
-    fun deleteRepository(@NotBlank @RequestParam id: Long): ResponseEntity<Any> {
+    fun deleteRepository(@CookieValue("jwt") jwt: String?, @NotBlank @RequestParam id: Long): ResponseEntity<Any> {
+        if (jwt.isNullOrBlank()) {
+            return ResponseEntity(MessageDTO("It is not authenticated. Please log in"), HttpStatus.UNAUTHORIZED)
+        }
+        val body = Jwts.parser().setSigningKey("secret".encodeToByteArray()).parseClaimsJws(jwt).body
+        if (body["role"] == "STUDENT") return ResponseEntity(
+            MessageDTO("You do not have permissions to access this resource"),
+            HttpStatus.UNAUTHORIZED
+        )
         repositoryService.deleteById(id)
         return ResponseEntity(MessageDTO("Repository has been deleted successfully"), HttpStatus.OK)
     }
@@ -203,7 +237,10 @@ class RepositoryController(private val repositoryService: RepositoryService) {
                 ]
             )]
     )
-    fun getAll(): ResponseEntity<List<Repository>> {
+    fun getAll(@CookieValue("jwt") jwt: String?): ResponseEntity<Any> {
+        if (jwt.isNullOrBlank()) {
+            return ResponseEntity(MessageDTO("It is not authenticated. Please log in"), HttpStatus.UNAUTHORIZED)
+        }
         return ResponseEntity(repositoryService.findByAll(), HttpStatus.OK)
     }
 }
